requirements.txt
customtkinter
tensorflow
numpy
pillow
opencv-python